move saveCompressedMask to SaveModule
